+++ src/js/NativeRNSentry.ts
@@ -33,6 +33,7 @@ export interface Spec extends TurboModule {
   setTag(key: string, value: string): void;
   enableNativeFramesTracking(): void;
   fetchModules(): Promise<string | undefined | null>;
+  fetchViewHierarchy(): Promise<number[] | undefined | null>;
 }
 
 export type NativeAppStartResponse = {